updated changes-see notes.
fixed auto id from jamming opposing players weapons in a logic loop.
autoid_on just reveals the player's name or their objects that they build.  The current code was a spy detector with "friendly/enemy" regardless whether the target viewed as a spy.  If the spy was undercover, they were revealed anyway. (smh). 
The objects built, like sentry guns and teslas are described as friendly or enemy.
The spy can run around with cutomskin as default or change colors.
The warning for pants color wrong as spy in disguise is for testing only now. will be removed.
the spy must be revealed when they fire, or you damage them and look for damage sound.
The engineer's security cam will stay as a spy detector. The tesla upgrade for spy detector is the same.
Considering adding a team message from your team medics alerting to a spy as theri netname.
Considering adding a change name to spy class to really muck thing up with identification.
Typically once a player starts a spy run/death or two, people figure it out that the player's name is likely a spy.
slight fix on the voting routines:
1. if you are a fresh new connected player spawn you get to vote almost immediately.
2. The "you have already voted." logic is missing some case scenarios.  In the case of voting to use an entity/switch/door your switch that you have voted is toggled back off.  That way you can request a vote again or vote on another vote already in progress (changemap, exit,etc..)
3. The colored dots for summoned monsters is not a neat matrix of the team number summoning the dead monster. It needs to reflect the color the team of the summoner. Currently dot3 and newdot are used if avaiable as models. A feature request for correct coloring forthcoming.
4. Numerous dialog fixes with plain text.
5. Gravity well does not pull in backpacks but will detonate nearby ones with emp if in range if created by dying monsters/players.


2.