updates to close issues. see notes.
all issues related to auto id, lag, spots for players of the color of their team is on. 
spots can be removed with invisibility, or spy feign death.
spots help in the case of customskins or changed teams since players can switch easily.
The downside is that auto ID always tells the player the true netname of the player, even a spy in disquise. 
the problem is that people know when one such player runs spy attacks several times in a row.
The alternative is for the enemy to not hog spy class but rotate once in a while to throw them off.
Another idea is that the reported netname can be confused if it's a spy undercover being looked at, but the point is moot once the observe sees the team color of the player.  (heh)