Added color spot fobs to follow players. Spy updt.
Since changing pants, pants color, top color and the use of custom skins, it's hard to identify teammates and enemies. Specifically since perfidy is allowed with the spy. Some changes were made to deal with auto_id on/off and identification were made. A colored spot sprite player fob that identify the team color follow the player. However they move to top of map and offset while invisible, feigning death as spy, or undercover with different skins as spy. The color spot also prevents all black or shadowed skins to prevent camping in maps with improper lighting or vis. Auto ID will tell you the names of players or objects built like the sentry or tesla. The netname of the spy is not revealed. However it is easy to deduce using this in tandem with the main scoreboard. It's behavior is not changed.
Corrected the amount of time for monster spawns to 15 seconds in survival mode, and progressively spawns faster on large monster maps so that the ending does not take 45-180 seconds to spawn, but halves it as it goes to the last 20 monsters.
Updated "cmd help misc" with a help  listing of existing commands for monster spawn time, total monsters, entities spawned, etc.
Specified variables unused for MarioWorld style of maps for Agent. Functions for map specific calls of the MarioWorld type of physics forthcoming. Reverted droptofloor() back to two parameter input to reduce compiler warnings. Still at 4(hook, 1 sky), and 7 unused variables(for now).