Fixed voting text messages/sound precache
The map name displayed for "cmd vote map" and "cmd votenext" displays notifications several times in console log and user screen. The centerprint advises the client to look there. This bug of centerprint not being able to print the current variable for map name is untenable due to the way it is stored in the "world" and pulled out and manipulated with local vars defined. Even when using a global string variable to pass the correct map name within memory blocks, it doesn't refresh correctly.