Cleaned up #ifdef, formatted if statements.
yep// all better
Cleaned up format of source some more.
"else if " is sloppy coding and leads to confusion when reading it.
"if (test_condition) { "  is another no no when there are nested If statements without proper indentation which leads to reading confusion.
Applied in several formats across multiple devs.  This makes it standard and easy reading.
Removed the nested #ifdef statements to reduce compilation time and also probably execution time of the the code when it loops around. 
There's no need to check 150 times if the code is mtf_coop, or mtfcoop_expsys, etc.. It's been that way for 5 years and nobody added a variation on this compilation.  This makes the server COOP only in respect of the engine and game mode switches in the server config files.  However, there is nothing stopping someone from loading up a DM deathmatch only map and just pick red/blue team for frag counting. Whether or not it has monsters in it.