fixed setrune, megahealth monster drops.
The error is erroneous, runes are set, it's just that there is no such thing as type 0 or before the timer is ready to do deploy a rune.
Also added monster dropped item mega health to actually help players when their health is over 250 instead of only below that. 
The players are going to need it when there is a room full of bad assed wraths or shamblers.