droptofloor fixed.
The function droptofloor() has no parameters after reviewing the source in server c++. So just bind it to #34 as a plain void call.
Made dispenser into a prox bomb. If a monster or enemy touches it to use use it, boom.
Monsters get pissed and attack each other more frequently.
"cmd tent" does remove things and lower 512 limit. But once it's passed, it's downhill from there. Testing.
Might add it to the loop for silent cleanup in hairy battles.